gameController almost done

+++ src/main/java/aDarbellay/s05/t1/model/player/CautiousPlayer.java
@@ -27,12 +27,12 @@ public class CautiousPlayer implements Player {
     }
 
     @Override
-    public long getId() {
+    public int getId() {
         return 0;
     }
 
     @Override
-    public void setId(long i) {
+    public void setId(int i) {
 
     }
